+++ miranda/protocols/Gadu-Gadu/libgadu/events.c
@@ -1705,7 +1705,7 @@ struct gg_event *gg_watch_fd(struct gg_session *sess)
 			l.uin = gg_fix32(sess->uin);
 			l.status = gg_fix32(sess->initial_status ? sess->initial_status : GG_STATUS_AVAIL);
 			l.version = gg_fix32(sess->protocol_version);
-			l.local_port = gg_fix16(gg_dcc_port);
+			l.local_port = gg_fix16((uint16_t)gg_dcc_port);
 			l.image_size = sess->image_size;
 
 			if (sess->external_addr && sess->external_port > 1023) {
